refactor: group exceptions in one file
