+++ lib/requests.dart
@@ -4,3 +4,4 @@ export 'src/requests.dart';
 export 'src/response.dart';
 export 'src/cookie.dart';
 export 'src/cookie_jar.dart';
+export 'src/exception.dart';
