--- lib/src/response_error.dart
@@ -1,8 +1,0 @@
-part of 'response.dart';
-
-class HTTPException implements Exception {
-  final String message;
-  final Response response;
-
-  HTTPException(this.message, this.response);
-}
